✨feat: adiciona página de listar produtos e outros refinamentos

+++ src/main/resources/schema-postgresql.sql
@@ -2,5 +2,5 @@ CREATE TABLE IF NOT EXISTS produto (
      id serial PRIMARY KEY,
      nome varchar(50),
      descricao varchar(200),
-     preco float
+     preco DOUBLE PRECISION
 );
